farm/temperature-sensor/requirements.txt
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
CounterFit==0.1.4.dev9
counterfit-connection==0.1.0.dev5
counterfit-shims-grove==0.1.4.dev5
dnspython==2.6.1
eventlet==0.33.3
Flask==2.1.2
Flask-SocketIO==5.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paho-mqtt==1.5.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.2.1
Werkzeug==2.0.0
wsproto==1.2.0
